push

+++ client/src/app/app.component.ts
@@ -1,22 +1,26 @@
 import { Component } from '@angular/core';
 import {Router} from '@angular/router';
+import {Http,Headers} from '@angular/http'
 import {ProfileService,EditProfile} from '../app/profile.service';
+import {LoginService} from '../app/login.service';
 
 
 import { Idle } from 'idlejs/dist';
 import { longStackSupport } from 'q';
+import { store } from '@angular/core/src/render3';
 
 declare var jquery:any;
 declare var $ :any;
 const idle = new Idle()
   .whenNotInteractive()
-  .within(5)
+  .within(10)
   .do(() => {
      console.log('IDLE');
      sessionStorage.clear();
-     console.log(sessionStorage.getItem('email')); 
+     console.log(sessionStorage.getItem('name')); 
      $('.logout').hide();     
      window.location.href = "/login";
+     //this.logout();
   })
   .start();
 
@@ -24,43 +28,43 @@ const idle = new Idle()
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
-  providers: [ProfileService,EditProfile],
+  providers: [ProfileService,EditProfile,LoginService],
 })
 export class AppComponent {
 
-  constructor(private router:Router,private profileService:ProfileService,private editProfile:EditProfile) { 
+  constructor(private router:Router,private profileService:ProfileService,private editProfile:EditProfile,private loginService:LoginService) { 
     this.profileService.setNameMethodCalled.subscribe((res)=>{
       this.user=res.name;
       this.id=res._id;
     })
   }
   title = 'client';
-  user=sessionStorage.getItem('email');
+  user=sessionStorage.getItem('name');
   id=sessionStorage.getItem('id');
   
-  
-
   logout(){
-    console.log('logged out');    
-    console.log(sessionStorage.getItem('email'));
-     //sessionStorage.setItem('email',"");
-         
-     this.user="";
-     sessionStorage.clear();
-     console.log(sessionStorage.getItem('email')); 
-     $('.logout').hide();
-     this.router.navigate(['login']);    
-     
-    }   
+    this.loginService.logout({"id":sessionStorage.getItem('id')})
+    .subscribe(res=>{
+      if(res.status==true){
+      console.log('logged out');    
+      console.log(sessionStorage.getItem('name'));            
+      this.user="";
+      sessionStorage.clear();   
+      console.log(sessionStorage.getItem('name')); 
+      $('.logout').hide();
+      this.router.navigate(['login']);        
+      }
+      else{
+       alert(res.message);
+      }
+    });    
+  }   
 
     profile(data){
       console.log(data);
       this.router.navigate(['registration']);
       this.profileService.getProfile(data).subscribe((res)=>{
         this.editProfile.editProfile(res);
-      })
-      
+      });      
     }
-
-
 }
